EA-6812: add lib64 symlink so PHP can find what it needs
For some reason it doesn't use `%{_lib}` and if we change it to do so who
knows what might break.

Even though RPM can doe snot handle symlink to dir and vice verse
changes its much simpler to just do the symlink for now `¯\_(ツ)_/¯` and
if we ever switch to `%{_lib}` deal with it then.